requirements.txt
requests==2.22.0
python-dotenv==0.21.0
Flask==2.2.5
Flask-Cors==3.0.9